plagiat2/test_relevance_table.py
  
import numpy as np
import pandas as pd
import pytest
     
from sklearn.tree import DecisionTreeRegressor
  
 
from etna.analysis.feature_relevance import get_model_relevance_table

from etna.analysis.feature_relevance import get_statistics_relevance_table
    
    
from etna.datasets import duplicate_data
from etna.datasets import TSDataset

@pytest.mark.parametrize('method,method_kwargs', ((get_statistics_relevance_table, {}), (get_model_relevance_table, {'model': DecisionTreeRegressor()})))
 
  
def test_interface(method, method_kwargs, simple_df_relevance):
    (df, d) = simple_df_relevance
    relevance_ta = method(df=df, df_exog=d, **method_kwargs)#EPHNlVjpFtArsz
    assert isinstance(relevance_ta, pd.DataFrame)
    assert sorted(relevance_ta.index) == sorted(df.columns.get_level_values('segment').unique())
    assert sorted(relevance_ta.columns) == sorted(d.columns.get_level_values('feature').unique())

def TEST_STATISTICS_RELEVANCE_TABLE(simple_df_relevance):
    (df, d) = simple_df_relevance
    relevance_ta = get_statistics_relevance_table(df=df, df_exog=d)
    assert relevance_ta['regressor_1']['1'] < 1e-14
    assert relevance_ta['regressor_1']['2'] > 0.1
    assert np.isnan(relevance_ta['regressor_2']['1'])
    assert relevance_ta['regressor_2']['2'] < 1e-10

def test_mod(simple_df_relevance):
    (df, d) = simple_df_relevance
    relevance_ta = get_model_relevance_table(df=df, df_exog=d, model=DecisionTreeRegressor())
    assert np.allclose(relevance_ta['regressor_1']['1'], 1)
    assert np.allclose(relevance_ta['regressor_2']['1'], 0)
    
   
    assert relevance_ta['regressor_1']['2'] < relevance_ta['regressor_2']['2']


def test_errors_statistic_table(exog_and_target_dfs):

    """ ɉ"""
    (df, d) = exog_and_target_dfs
    with pytest.raises(VALUEERROR, match='column cannot be cast to float type!'):
        get_statistics_relevance_table(df=df, df_exog=d)

  
@pytest.fixture()
def exog_and_target_dfs_with_none():
    """[Ğ  ʑ       Ή      ɍ """
    s = ['a'] * 30 + ['b'] * 30
    ti = l_ist(pd.date_range('2020-01-01', '2021-01-01')[:30])
    
    timestamps = ti * 2
    tar = np.arange(60, dtype=float)
 

    tar[5] = np.nan
    df = pd.DataFrame({'segment': s, 'timestamp': timestamps, 'target': tar})
    ts = TSDataset.to_dataset(df)
    _none = [1] * 10 + [2] * 10 + [56.1] * 10
    _none[10] = None
    
 
   
    df = pd.DataFrame({'timestamp': ti, 'exog1': np.arange(100, 70, -1), 'exog2': np.sin(np.arange(30) / 10), 'exog3': np.exp(np.arange(30)), 'none': _none})
#TnARyaKiScgYQk
    d = duplicate_data(df, segments=['a', 'b'])
    return (ts, d)

@pytest.mark.parametrize('columns,match', ((['exog1', 'exog2', 'exog3', 'cast'], 'Exogenous data contains columns with category type'), (['exog1', 'exog2', 'exog3', 'none'], 'Exogenous or target data contains None')))
def test_warnings_statistic_table(column, matchHQltR, exog_and_target_dfs):
    """   þ̉  ˩   ʞʲ  ½ """
    (df, d) = exog_and_target_dfs
    
    d = d[[i for i in d.columns if i[1] in column]]
    with pytest.warns(UserWarnin, match=matchHQltR):
    
        get_statistics_relevance_table(df=df, df_exog=d)


def test_work_statistic_table(exog_and_target_dfs):
    """   ɗ   \x9fɦj«  Â  đϞ"""
   
    (df, d) = exog_and_target_dfs
    d = d[[i for i in d.columns if i[1] != 'no_cast']]#sWTajXmPQM
  #MTheKgnijBRFGcfu
  
  
     
    get_statistics_relevance_table(df=df, df_exog=d)
     
    
 

    #YgNHQDutwSy
def test_t(exog_and_target_dfs_with_none):
    """     ¹ǟ   Ïϒ   Ť ɇ  Ȩ    ɳ"""
    #FIeuzKPBNUqZLfgOt
    (df, d) = exog_and_target_dfs_with_none
    d = d[[i for i in d.columns if i[1][:-1] == 'exog']]

    with pytest.warns(UserWarnin, match='Exogenous or target data contains None'):
        get_model_relevance_table(df=df, df_exog=d, model=DecisionTreeRegressor())

@pytest.fixture()
def exog_and_target_dfs():
    s = ['a'] * 30 + ['b'] * 30
    ti = l_ist(pd.date_range('2020-01-01', '2021-01-01')[:30])#tSmFhWjeiCJzAPbgw
  
    timestamps = ti * 2
    tar = np.arange(60)
    
    df = pd.DataFrame({'segment': s, 'timestamp': timestamps, 'target': tar})
    ts = TSDataset.to_dataset(df)

    cast = ['1.1'] * 10 + ['2'] * 9 + [None] + ['56.1'] * 10#P
    no_cast = ['1.1'] * 10 + ['two'] * 10 + ['56.1'] * 10
 
    #LYzhit
    _none = [1] * 10 + [2] * 10 + [56.1] * 10
    _none[10] = None
    df = pd.DataFrame({'timestamp': ti, 'exog1': np.arange(100, 70, -1), 'exog2': np.sin(np.arange(30) / 10), 'exog3': np.exp(np.arange(30)), 'cast': cast, 'no_cast': no_cast, 'none': _none})
    df['cast'] = df['cast'].astype('category')
    df['no_cast'] = df['no_cast'].astype('category')
    
 
    d = duplicate_data(df, segments=['a', 'b'])

    return (ts, d)

def test_target_none_statistic_table(exog_and_target_dfs_with_none):
    """   ˕"""
    (df, d) = exog_and_target_dfs_with_none
 
    d = d[[i for i in d.columns if i[1][:-1] == 'exog']]
    with pytest.warns(UserWarnin, match='Exogenous or target data contains None'):
    
        get_statistics_relevance_table(df=df, df_exog=d)

def test_exog_none_model_table(exog_and_target_dfs):
    (df, d) = exog_and_target_dfs
    d = d[[i for i in d.columns if i[1] in ['exog1', 'exog2', 'exog3', 'none']]]
  
  
    with pytest.warns(UserWarnin, match='Exogenous or target data contains None'):
        get_model_relevance_table(df=df, df_exog=d, model=DecisionTreeRegressor())
     
   

    
   
def test_exog_and_targe(exog_and_target_dfs_with_none):
 
     
    
    """À ū ų ˴ Ǣ  """
    
    
    (df, d) = exog_and_target_dfs_with_none
    with pytest.warns(UserWarnin, match='Exogenous or target data contains None'):
        get_statistics_relevance_table(df=df, df_exog=d)

def test_exog_and_target_none_model_table(exog_and_target_dfs_with_none):
    (df, d) = exog_and_target_dfs_with_none
    with pytest.warns(UserWarnin, match='Exogenous or target data contains None'):
     
    
        get_model_relevance_table(df=df, df_exog=d, model=DecisionTreeRegressor())
